Fix charge on BGP (32 bit error). Each group writes uk.%g000.dat for multiple inputs.
git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@5360 e5b18d87-469d-4833-9cc0-8cdfa06e9491

+++ src/ParticleIO/ESHDFParticleParser.cpp
@@ -134,8 +134,8 @@ namespace qmcplusplus {
       a.read(g,"name");
       int ii=tspecies.addSpecies(aname);
 
-      double q=-1;
-      HDFAttribIO<double> b(q);
+      int q=-1;
+      HDFAttribIO<int> b(q);
       b.read(g,charge_tag.c_str());
       tspecies(icharge,ii)=q;
